apps/crawler/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright==1.47.0
google-generativeai==0.8.3
supabase==2.7.4
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
